+++ changelog.md
@@ -5,6 +5,7 @@
 * Store Tagify instance on the original node for referencing later as .data('tagify-instance')
 * Adds a reset method for reverting to an empty state
 * Throw exceptions if the specified method does not exist
+* invalidCb option to run a callback when an invalid tag is found. It also provides a hook to fixup / force tagify to add the invalid tag.
 
 0.0.1
 =====